requirements.txt
deepspeed==0.14.4
geomloss==0.2.6
matplotlib==3.10.6
natsort==8.4.0
numpy==2.3.3
pandas==2.3.2
pytorch_lightning==2.3.0
scanpy==1.11.4
scikit_learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
torch==2.6.0+cu118
umap_learn==0.5.7
